Add minimal support for arbitrary characters
This change adds minimum support for displaying arbitrary characters that are not defined
in the character list. For any character that we try to animate to/from that we don't have
an index for, it will simply assume that the index is at the beginning of the list and
animate accordingly.